template_dir: Various changes.
Don't allow relative paths for SDK installation. (The files are first
extracted to tmp folder, so the relative path would be relative to
tmp, not the cwd)
Prompt to remove any existing files at the SDK install folder if they exist.
This can be automatic "yes" if specified -y on the commandline

Remove site-config* and environment-setup* files after SDK installation.
These files are deemed confusing/redundunt for Zephyr users.
However, they can be preserved if -nocleanup is specified on the command
line.

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>